# openapi-generator/spec/openapi-petstore.yaml
openapi: 3.0.2
info:
    title: Swagger Petstore - OpenAPI 3.0
    description: >-
        This is a sample Pet Store Server based on the OpenAPI 3.0
        specification.  You can find out more about

        Swagger at [http://swagger.io](http://swagger.io). In the third iteration
        of the pet store, we've switched to the design first approach!

        You can now help us improve the API whether it's by making changes to the
        definition itself or to the code.

        That way, with time, we can improve the API in general, and expose some
        of the new features in OAS3.



        Some useful links:

        - [The Pet Store
        repository](https://github.com/swagger-api/swagger-petstore)

        - [The source API definition for the Pet
        Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
    termsOfService: 'http://swagger.io/terms/'
    contact:
        email: apiteam@swagger.io
    license:
        name: Apache 2.0
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
    version: 1.0.17
...
servers:
    - 
        url: /api/v3
...
paths:
...
    /user:
        post:
            tags:
                - user
            summary: Create user
            description: This can only be done by the logged in user.
            operationId: createUser
            requestBody:
                description: Created user object
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/User'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                default:
                    description: successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/User'
    /user/createWithList:
        post:
            tags:
                - user
            summary: Creates list of users with given input array
            description: Creates list of users with given input array
            operationId: createUsersWithListInput
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                            items:
                                $ref: '#/components/schemas/User'
            responses:
                '200':
                    description: Successful operation
                    content:
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/User'
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                default:
                    description: successful operation
    /user/login:
        get:
            tags:
                - user
            summary: Logs user into the system
            description: ''
            operationId: loginUser
            parameters:
                - 
                    name: username
                    in: query
                    description: The user name for login
                    required: false
                    schema:
                        type: string
                - 
                    name: password
                    in: query
                    description: The password for login in clear text
                    required: false
                    schema:
                        type: string
            responses:
                '200':
                    description: successful operation
                    headers:
                        X-Rate-Limit:
                            description: calls per hour allowed by the user
                            schema:
                                type: integer
                                format: int32
                        X-Expires-After:
                            description: date in UTC when token expires
                            schema:
                                type: string
                                format: date-time
                    content:
                        application/xml:
                            schema:
                                type: string
                        application/json:
                            schema:
                                type: string
                '400':
                    description: Invalid username/password supplied
...
    '/user/{username}':
        get:
            tags:
                - user
            summary: Get user by user name
            description: ''
            operationId: getUserByName
            parameters:
                - 
                    name: username
                    in: path
                    description: 'The name that needs to be fetched. Use user1 for testing. '
                    required: true
                    schema:
                        type: string
            responses:
                '200':
                    description: successful operation
                    content:
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/User'
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: Invalid username supplied
                '404':
                    description: User not found
        put:
            tags:
                - user
            summary: Update user
            description: This can only be done by the logged in user.
            operationId: updateUser
            parameters:
                - 
                    name: username
                    in: path
                    description: name that need to be deleted
                    required: true
                    schema:
                        type: string
            requestBody:
                description: Update an existent user in the store
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/User'
                    application/x-www-form-urlencoded:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                default:
                    description: successful operation
        delete:
            tags:
                - user
            summary: Delete user
            description: This can only be done by the logged in user.
            operationId: deleteUser
            parameters:
                - 
                    name: username
                    in: path
                    description: The name that needs to be deleted
                    required: true
                    schema:
                        type: string
            responses:
                '400':
                    description: Invalid username supplied
                '404':
                    description: User not found
components:
    schemas:
...
        User:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                    example: 10
                username:
                    type: string
                    example: theUser
                firstName:
                    type: string
                    example: John
                lastName:
                    type: string
                    example: James
                email:
                    type: string
                    example: john@email.com
                password:
                    type: string
                    example: '12345'
                phone:
                    type: string
                    example: '12345'
                userStatus:
                    type: integer
                    description: User Status
                    format: int32
                    example: 1
            xml:
                name: user
...
    securitySchemes:
        petstore_auth:
            type: oauth2
            flows:
                implicit:
                    authorizationUrl: 'https://petstore3.swagger.io/oauth/authorize'
                    scopes:
                        'write:pets': modify pets in your account
                        'read:pets': read your pets
        api_key:
            type: apiKey
            name: api_key
            in: header